seperate files for slash commands
added slash commands models and functions in new files
